+++ src/arch/armv8/inc/arch/vgic.h
@@ -103,7 +103,7 @@ enum vgic_reg_handler_info_id {
 struct vgic_reg_handler_info {
     void (*reg_access)(emul_access_t *, struct vgic_reg_handler_info *,
                        bool gicr_accces, uint64_t vgicr_id);
-    uint8_t alignment;
+    size_t alignment;
     size_t regid;
     uintptr_t regroup_base;
     size_t field_width;
@@ -123,7 +123,7 @@ void vgic_yield_ownership(vcpu_t *vcpu, vgic_int_t *interrupt);
 void vgic_emul_generic_access(emul_access_t *, struct vgic_reg_handler_info *,
                               bool, uint64_t);
 void vgic_send_sgi_msg(vcpu_t *vcpu, uint64_t pcpu_mask, uint64_t int_id);
-uint64_t vgic_get_itln(const struct gic_dscrp *gic_dscrp);
+size_t vgic_get_itln(const struct gic_dscrp *gic_dscrp);
 
 /* interface for version specific vgic */
 bool vgic_int_vcpu_is_target(vcpu_t *vcpu, vgic_int_t *interrupt);
