+++ src/arch/armv8/iommu.c
@@ -28,9 +28,9 @@ int iommu_arch_init()
     return -1;
 }
 
-static int32_t iommu_vm_arch_init_ctx(vm_t *vm)
+static ssize_t iommu_vm_arch_init_ctx(vm_t *vm)
 {
-    int32_t ctx_id = vm->iommu.arch.ctx_id;
+    ssize_t ctx_id = vm->iommu.arch.ctx_id;
     if (ctx_id < 0) {
 
         /* Set up ctx bank to vm address space in an available ctx. */
@@ -51,7 +51,7 @@ static int32_t iommu_vm_arch_init_ctx(vm_t *vm)
 
 static int iommu_vm_arch_add(vm_t *vm, uint16_t mask, uint16_t id)
 {
-    int32_t vm_ctx = iommu_vm_arch_init_ctx(vm);
+    ssize_t vm_ctx = iommu_vm_arch_init_ctx(vm);
     uint16_t glbl_mask = vm->iommu.arch.global_mask;
     uint16_t prep_mask = (mask & SMMU_ID_MSK) | glbl_mask;
     uint16_t prep_id = (id & SMMU_ID_MSK);
@@ -62,7 +62,7 @@ static int iommu_vm_arch_add(vm_t *vm, uint16_t mask, uint16_t id)
     }
 
     if (!smmu_compatible_sme_exists(prep_mask, prep_id, vm_ctx, group)) {
-        int32_t sme = smmu_alloc_sme();
+        size_t sme = smmu_alloc_sme();
         if(sme < 0){
             INFO("iommu: smmuv2 no more free sme available.");
             return -1;
@@ -86,7 +86,7 @@ int iommu_arch_vm_init(vm_t *vm, const vm_config_t *config)
     vm->iommu.arch.ctx_id = -1;
 
     /* This section relates only to arm's iommu so we parse it here. */
-    for (int i = 0; i < config->platform.arch.smmu.group_num; i++) {
+    for (size_t i = 0; i < config->platform.arch.smmu.group_num; i++) {
         /* Register each group. */
         const struct smmu_group *group =
             &config->platform.arch.smmu.smmu_groups[i];
